+++ src/plugins/validate-ajv/index.ts
@@ -12,15 +12,35 @@ import type {
 import { wrappedValidateStorageFactory } from '../../plugin-helpers.ts';
 
 
-const ajv = new Ajv({
-    strict: false
-});
+let ajv: Ajv;
+
+
+export function getAjv() {
+    if (!ajv) {
+        ajv = new Ajv({
+            strict: true
+        });
+        ajv.addKeyword('version');
+        ajv.addKeyword('keyCompression');
+        ajv.addKeyword('primaryKey');
+        ajv.addKeyword('indexes');
+        ajv.addKeyword('encrypted');
+        ajv.addKeyword('final');
+        ajv.addKeyword('sharding');
+        ajv.addKeyword('internalIndexes');
+        ajv.addKeyword('attachments');
+        ajv.addKeyword('ref');
+        ajv.addKeyword('crdt');
+    }
+    return ajv;
+}
+
 
 
 export function getValidator(
     schema: RxJsonSchema<any>
 ) {
-    const validator = ajv.compile(schema);
+    const validator = getAjv().compile(schema);
     return (docData: RxDocumentData<any>) => {
         const isValid = validator(docData);
         if (isValid) {
@@ -30,6 +50,8 @@ export function getValidator(
         }
     };
 }
+
+
 
 export const wrappedValidateAjvStorage = wrappedValidateStorageFactory(
     getValidator,